further refinement
