+++ python/semantic_kernel/semantic_functions/function_call_response_template.py
@@ -1,0 +1,16 @@
+import json
+from dataclasses import dataclass
+from typing import Dict
+
+
+@dataclass
+class FunctionCallResponseTemplate:
+    name: str
+    arguments: str
+
+    @property
+    def parsed_arguments(self) -> Dict[str, str]:
+        try:
+            return json.loads(self.arguments)
+        except json.JSONDecodeError:
+            return {}
