+++ python/semantic_kernel/skill_definition/sk_function_context_parameter_decorator.py
@@ -2,7 +2,12 @@
 
 
 def sk_function_context_parameter(
-    *, name: str, description: str, default_value: str = ""
+    *,
+    name: str,
+    description: str,
+    default_value: str = "",
+    type: str = "string",
+    required: bool = False
 ):
     """
     Decorator for SK function context parameters.
@@ -11,6 +16,9 @@ def sk_function_context_parameter(
         name -- The name of the context parameter
         description -- The description of the context parameter
         default_value -- The default value of the context parameter
+        type -- The type of the context parameter
+        required -- Whether the context parameter is required
+
     """
 
     def decorator(func):
@@ -22,6 +30,8 @@ def sk_function_context_parameter(
                 "name": name,
                 "description": description,
                 "default_value": default_value,
+                "type": type,
+                "required": required,
             }
         )
         return func
